migrate photo lib access client
