+++ WegliKit/Sources/PhotoLibraryAccessClient/Mock.swift
@@ -1,0 +1,10 @@
+import ComposableArchitecture
+import Photos
+
+public extension PhotoLibraryAccessClient {
+  static let noop = Self(
+    requestAuthorization: { .none },
+    authorizationStatus: { .notDetermined }
+  )  
+}
+
